[JENKINS-32651] Fix calculation of report begin date
Currently, computation is done using Calendar.roll() which is wrong.
roll() only changes a specific unit and doesn't update larger fields.
For example, rolling back 2 week on the first week of a month sets
the calendar week to the last week of the month, which is in the future.
Obviously, when the begin date is in the future, there's nothing in the
report.

Using Calendar.add() with a negative value substracts the right amount
of time from the current time.